requirements.txt
langchain>=0.1.16
langchain-core>=0.1.37
langchain-community>=0.0.30
langchain-openai>=0.1.6
chromadb>=0.4.24
sentence-transformers>=2.2.2
pypdf>=3.16.4
pyyaml>=6.0.1
tqdm>=4.66.2
cmake>=3.25
beautifulsoup4>=4.13.4
llama-cpp-python